bg thread for computations

+++ field.cc
@@ -2,7 +2,7 @@
 
 namespace miner {
 
-void field::reset ( size_t rows, size_t cols ) {
+void field::reset ( int rows, int cols ) {
     mines_nr_ = 0;
     rows_ = rows;
     cols_ = cols;
@@ -12,7 +12,7 @@ void field::reset ( size_t rows, size_t cols ) {
 
 
 void field::set_mine ( coord c, bool v ) {
-    size_t idx = c.row * cols_ + c.col;
+    int idx = c.row * cols_ + c.col;
     if ( data_[idx] ) {
 	if ( !v ) {
 	    data_[idx] = false;
@@ -28,7 +28,7 @@ void field::set_mine ( coord c, bool v ) {
 }
 
 
-void field::gen_random ( size_t rows, size_t cols, size_t mines_nr ) {
+void field::gen_random ( int rows, int cols, int mines_nr ) {
     srand48(time(nullptr));
     reset(rows, cols);
     
@@ -37,7 +37,7 @@ void field::gen_random ( size_t rows, size_t cols, size_t mines_nr ) {
     
     mines_nr_ = mines_nr;
     while(mines_nr) {
-	coord c{size_t(drand48() * rows_), size_t(drand48() * cols_)};
+	coord c{int(drand48() * rows_), int(drand48() * cols_)};
 	if ( !is_mine(c) ) {
 	    data_[c.row * rows_ + c.col] = true;
 	    --mines_nr;
@@ -46,8 +46,8 @@ void field::gen_random ( size_t rows, size_t cols, size_t mines_nr ) {
 }
 
 
-size_t field::nearby_mines_nr ( coord c ) const {
-    size_t rv{};
+int field::nearby_mines_nr ( coord c ) const {
+    int rv{};
     
     if ( c.row > 0 ) {
 	if ( c.col > 0 ) rv += is_mine({c.row-1, c.col-1});
